Extract directory checksum validator so we can audit uncataloged moabs
Fixes #2396
