+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config --auto-gen-only-exclude`
-# on 2023-09-11 16:55:28 UTC using RuboCop version 1.56.3.
+# on 2025-04-02 16:45:34 UTC using RuboCop version 1.75.1.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -8,7 +8,7 @@
 
 # Offense count: 1
 # This cop supports safe autocorrection (--autocorrect).
-# Configuration parameters: Max, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, AllowedPatterns.
+# Configuration parameters: Max, AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, AllowedPatterns, SplitStrings.
 # URISchemes: http, https
 Layout/LineLength:
   Exclude:
@@ -30,7 +30,7 @@ Lint/MissingSuper:
   Exclude:
     - 'config/initializers/okcomputer.rb'
 
-# Offense count: 21
+# Offense count: 22
 # Configuration parameters: AllowedMethods, AllowedPatterns, CountRepeatedAttributes, Max.
 Metrics/AbcSize:
   Exclude:
@@ -40,7 +40,7 @@ Metrics/AbcSize:
     - 'app/models/preserved_object.rb'
     - 'app/models/replication/druid_version_zip.rb'
     - 'app/services/audit/catalog_to_moab.rb'
-    - 'app/services/audit/checksum_validator.rb'
+    - 'app/services/audit/checksum_validation_orchestrator.rb'
     - 'app/services/audit/checksum_validator_utils.rb'
     - 'app/services/audit/replication_support.rb'
     - 'app/services/audit/replication_to_endpoint_base.rb'
@@ -57,7 +57,7 @@ Metrics/ClassLength:
     - 'app/models/replication/druid_version_zip.rb'
     - 'app/services/moab_record_service/base.rb'
 
-# Offense count: 36
+# Offense count: 38
 # Configuration parameters: CountComments, CountAsOne, AllowedMethods, AllowedPatterns.
 Metrics/MethodLength:
   Max: 33
@@ -65,7 +65,7 @@ Metrics/MethodLength:
 # Offense count: 3
 # Configuration parameters: EnforcedStyle, CheckMethodNames, CheckSymbols, AllowedIdentifiers, AllowedPatterns.
 # SupportedStyles: snake_case, normalcase, non_integer
-# AllowedIdentifiers: capture3, iso8601, rfc1123_date, rfc822, rfc2822, rfc3339, x86_64
+# AllowedIdentifiers: TLS1_1, TLS1_2, capture3, iso8601, rfc1123_date, rfc822, rfc2822, rfc3339, x86_64
 Naming/VariableNumber:
   Exclude:
     - 'spec/services/moab_storage_root_report_service_spec.rb'
@@ -81,7 +81,7 @@ RSpec/InstanceVariable:
   Exclude:
     - 'spec/services/moab_on_storage/storage_directory_spec.rb'
 
-# Offense count: 367
+# Offense count: 368
 # Configuration parameters: AllowSubject.
 RSpec/MultipleMemoizedHelpers:
   Max: 21
@@ -90,13 +90,13 @@ RSpec/MultipleMemoizedHelpers:
 RSpec/StubbedMock:
   Exclude:
     - 'spec/services/audit/catalog_to_moab_spec.rb'
-    - 'spec/services/audit/checksum_validator_spec.rb'
+    - 'spec/services/audit/checksum_validation_orchestrator_spec.rb'
     - 'spec/services/audit/shared_examples_replication_to_endpoint.rb'
     - 'spec/services/moab_record_service/check_existence_spec.rb'
     - 'spec/services/moab_record_service/create_after_validation_spec.rb'
     - 'spec/services/moab_record_service/update_version_after_validation_spec.rb'
 
-# Offense count: 18
+# Offense count: 17
 # This cop supports safe autocorrection (--autocorrect).
 # Configuration parameters: EnforcedStyle.
 # SupportedStyles: slashes, arguments
@@ -139,14 +139,6 @@ Style/Documentation:
     - 'app/helpers/application_helper.rb'
     - 'app/models/application_record.rb'
     - 'config/application.rb'
-
-# Offense count: 435
-# This cop supports safe autocorrection (--autocorrect).
-# Configuration parameters: EnforcedStyle, EnforcedShorthandSyntax, UseHashRocketsWithSymbolValues, PreferHashRocketsForNonAlnumEndingSymbols.
-# SupportedStyles: ruby19, hash_rockets, no_mixed_keys, ruby19_no_mixed_keys
-# SupportedShorthandSyntax: always, never, either, consistent
-Style/HashSyntax:
-  Enabled: false
 
 # Offense count: 1
 # Configuration parameters: AllowedMethods.
