+++ app/jobs/audit/checksum_validation_job.rb
@@ -2,7 +2,7 @@
 
 module Audit
   # Confirm checksum for one Moab on storage and update MoabRecord in database
-  # @see Audit::ChecksumValidator
+  # @see Audit::ChecksumValidationOrchestrator
   class ChecksumValidationJob < ApplicationJob
     queue_as :checksum_validation
 
@@ -14,7 +14,7 @@ module Audit
 
     # @param [MoabRecord] moab_record object to checksum
     def perform(moab_record)
-      Audit::ChecksumValidator.new(moab_record).validate_checksums
+      Audit::ChecksumValidationOrchestrator.new(moab_record).validate_checksums
     end
   end
 end
